+++ library/src/main/java/de/mrapp/android/tabswitcher/layout/phone/PreviewDataBinder.java
@@ -22,9 +22,11 @@ import android.support.v4.util.Pair;
 import android.view.View;
 import android.view.View.MeasureSpec;
 import android.view.ViewGroup;
+import android.view.animation.AccelerateDecelerateInterpolator;
 import android.widget.ImageView;
 
 import de.mrapp.android.tabswitcher.Tab;
+import de.mrapp.android.tabswitcher.model.Model;
 import de.mrapp.android.tabswitcher.model.TabItem;
 import de.mrapp.android.util.multithreading.AbstractDataBinder;
 import de.mrapp.android.util.view.ViewRecycler;
@@ -51,6 +53,11 @@ public class PreviewDataBinder extends AbstractDataBinder<Bitmap, Tab, ImageView
     private final ViewRecycler<Tab, Void> contentViewRecycler;
 
     /**
+     * The model of the tab switcher, the tabs belong to.
+     */
+    private final Model model;
+
+    /**
      * Creates a new data binder, which allows to asynchronously render preview images of tabs and
      * display them afterwards.
      *
@@ -61,14 +68,19 @@ public class PreviewDataBinder extends AbstractDataBinder<Bitmap, Tab, ImageView
      *         The view recycler, which should be used to inflate the views, which are associated
      *         with tabs, as an instance of the class ViewRecycler. The view recycler may not be
      *         null
+     * @param model
+     *         The model of the tab switcher, the tabs belong to, as an instance of the type {@link
+     *         Model}. The model may not be null
      */
     public PreviewDataBinder(@NonNull final ViewGroup parent,
-                             @NonNull final ViewRecycler<Tab, Void> contentViewRecycler) {
-        super(parent.getContext(), new LruCache<Tab, Bitmap>(7));
+                             @NonNull final ViewRecycler<Tab, Void> contentViewRecycler,
+                             @NonNull final Model model) {
+        super(parent.getContext().getApplicationContext(), new LruCache<Tab, Bitmap>(7));
         ensureNotNull(parent, "The parent may not be null");
         ensureNotNull(contentViewRecycler, "The content view recycler may not be null");
         this.parent = parent;
         this.contentViewRecycler = contentViewRecycler;
+        this.model = model;
     }
 
     @Override
@@ -110,8 +122,22 @@ public class PreviewDataBinder extends AbstractDataBinder<Bitmap, Tab, ImageView
 
     @Override
     protected final void onPostExecute(@NonNull final ImageView view, @Nullable final Bitmap data,
-                                       @NonNull final TabItem... params) {
+                                       final long duration, @NonNull final TabItem... params) {
         view.setImageBitmap(data);
+
+        if (data != null) {
+            boolean useFadeAnimation = duration > model.getTabPreviewFadeThreshold();
+            view.setAlpha(useFadeAnimation ? 0f : 1f);
+            view.setVisibility(View.VISIBLE);
+
+            if (useFadeAnimation) {
+                view.animate().alpha(1f).setDuration(model.getTabPreviewFadeDuration())
+                        .setInterpolator(new AccelerateDecelerateInterpolator()).start();
+            }
+        } else {
+            view.setVisibility(View.INVISIBLE);
+        }
+
         view.setVisibility(data != null ? View.VISIBLE : View.GONE);
         TabItem tabItem = params[0];
         contentViewRecycler.remove(tabItem.getTab());